+++ knightbus-ui-console/KnightBus.UI.Console/Program.cs
@@ -1,11 +1,13 @@
 ﻿// See https://aka.ms/new-console-template for more information
 using KnightBus.UI.Console;
+using KnightBus.UI.Console.Providers;
+using KnightBus.UI.Console.Providers.ServiceBus;
 using Terminal.Gui;
 
 var connection = "";
 
 Application.Init();
-Application.Run(new MainWindow(new IQueueManager[] { new ServiceBusQueueManager(connection), new ServiceBusTopicManager(connection) }));
+Application.Run(new MainWindow(new IQueueManager[] { new QueueManager(connection), new TopicManager(connection) }));
