+++ knightbus-ui-console/KnightBus.UI.Console/Providers/IQueueManager.cs
@@ -1,0 +1,12 @@
+﻿using Azure.Messaging.ServiceBus;
+
+namespace KnightBus.UI.Console.Providers;
+
+public interface IQueueManager
+{
+    IEnumerable<QueueProperties> List(CancellationToken ct);
+    Task<QueueProperties> Get(string path, CancellationToken ct);
+    Task Delete(string path, CancellationToken ct);
+    Task<IReadOnlyList<ServiceBusReceivedMessage>> PeekDeadLetter(string name, int count, CancellationToken ct);
+    Task<int> MoveDeadLetters(string name, int count, CancellationToken ct);
+}
